Rewrite for applying new template
