+++ ros/src/computing/planning/motion/packages/waypoint_follower/CMakeLists.txt
@@ -57,7 +57,11 @@ add_library(libwaypoint_follower lib/libwaypoint_follower.cpp)
 add_dependencies(libwaypoint_follower
 waypoint_follower_generate_messages_cpp)
 
-add_executable(pure_pursuit nodes/pure_pursuit/pure_pursuit.cpp nodes/pure_pursuit/pure_pursuit_core.cpp nodes/pure_pursuit/pure_pursuit_viz.cpp)
+add_executable(pure_pursuit
+  nodes/pure_pursuit/pure_pursuit_node.cpp
+  nodes/pure_pursuit/pure_pursuit.cpp
+  nodes/pure_pursuit/pure_pursuit_core.cpp
+  nodes/pure_pursuit/pure_pursuit_viz.cpp)
 target_link_libraries(pure_pursuit libwaypoint_follower ${catkin_LIBRARIES})
 add_dependencies(pure_pursuit 
 waypoint_follower_generate_messages_cpp
