+++ ros/src/computing/planning/motion/packages/waypoint_follower/nodes/pure_pursuit/pure_pursuit.h
@@ -28,52 +28,83 @@
  *  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-#ifndef PURE_PURSUIT_CORE_H
-#define PURE_PURSUIT_CORE_H
+#ifndef PURE_PURSUIT_H
+#define PURE_PURSUIT_H
 
 // ROS includes
 #include <ros/ros.h>
-#include <geometry_msgs/TwistStamped.h>
 #include <geometry_msgs/PoseStamped.h>
-#include "runtime_manager/ConfigWaypointFollower.h"
-#include <visualization_msgs/Marker.h>
-// C++ includes
-#include <memory>
+#include <geometry_msgs/TwistStamped.h>
+
+// User defined includes
+#include "waypoint_follower/lane.h"
 #include "waypoint_follower/libwaypoint_follower.h"
+#include "runtime_manager/ConfigWaypointFollower.h"
 
-
+namespace waypoint_follower
+{
 enum class Mode
 {
   waypoint,
   dialog
 };
 
-namespace waypoint_follower
-{
 class PurePursuit
 {
-private:
+public:
+  PurePursuit();
+  ~PurePursuit();
 
-  //constant
+  // for getting data from ROS
+  void getConfigForROS(const runtime_manager::ConfigWaypointFollowerConstPtr &config);
+  void getCurrentPoseForROS(const geometry_msgs::PoseStampedConstPtr &msg);
+  void getCurrentVelocityForROS(const geometry_msgs::TwistStampedConstPtr &msg);
+  void getWayPointsForROS(const waypoint_follower::laneConstPtr &msg);
+  void getLinearInterpolationParameter(const bool &param);
+
+  // for debug on ROS
+  geometry_msgs::Point getPoseOfNextWaypoint() const
+  {
+    return current_waypoints_.getWaypointPosition(next_waypoint_number_);
+  }
+  geometry_msgs::Point getPoseOfNextTarget() const
+  {
+    return next_target_position_;
+  }
+  geometry_msgs::Pose getCurrentPose() const
+  {
+    return current_pose_.pose;
+  }
+
+  double getLookaheadDistance() const
+  {
+    return lookahead_distance_;
+  }
+  // processing
+  geometry_msgs::TwistStamped go();
+
+private:
+  // constant
   const double RADIUS_MAX_;
   const double KAPPA_MIN_;
 
-  bool linear_interpolate_;
+  // variables
+  bool is_linear_interpolation_;
 
   // config topic
-  int param_flag_;              // 0 = waypoint, 1 = Dialog
+  int param_flag_;                   // 0 = waypoint, 1 = Dialog
   double const_lookahead_distance_;  // meter
-  double initial_velocity_;     // km/h
+  double initial_velocity_;          // km/h
   double lookahead_distance_calc_ratio_;
   double minimum_lookahead_distance_;  // the next waypoint must be outside of this threshold.
   double displacement_threshold_;
   double relative_angle_threshold_;
 
-  bool waypoint_set_;
-  bool pose_set_;
-  bool velocity_set_;
-  int num_of_next_waypoint_;
-  geometry_msgs::Point position_of_next_target_;
+  bool is_waypoint_set_;
+  bool is_pose_set_;
+  bool is_velocity_set_;
+  int next_waypoint_number_;
+  geometry_msgs::Point next_target_position_;
   double lookahead_distance_;
 
   geometry_msgs::PoseStamped current_pose_;
@@ -90,57 +121,7 @@ private:
   void getNextWaypoint();
   geometry_msgs::TwistStamped outputZero() const;
   geometry_msgs::TwistStamped outputTwist(geometry_msgs::Twist t) const;
-
-public:
-  PurePursuit(bool linear_interpolate_mode)
-    : RADIUS_MAX_(9e10)
-    , KAPPA_MIN_(1/RADIUS_MAX_)
-    , linear_interpolate_(linear_interpolate_mode)
-    , param_flag_(0)
-    , const_lookahead_distance_(4.0)
-    , initial_velocity_(5.0)
-    , lookahead_distance_calc_ratio_(2.0)
-    , minimum_lookahead_distance_(6.0)
-    , displacement_threshold_(0.2)
-    , relative_angle_threshold_(10)
-    , waypoint_set_(false)
-    , pose_set_(false)
-    , velocity_set_(false)
-    , num_of_next_waypoint_(-1)
-    , lookahead_distance_(0)
-  {
-  }
-  ~PurePursuit()
-  {
-  }
-
-  // for ROS
-  void callbackFromConfig(const runtime_manager::ConfigWaypointFollowerConstPtr &config);
-  void callbackFromCurrentPose(const geometry_msgs::PoseStampedConstPtr &msg);
-  void callbackFromCurrentVelocity(const geometry_msgs::TwistStampedConstPtr &msg);
-  void callbackFromWayPoints(const waypoint_follower::laneConstPtr &msg);
-
-  // for debug
-  geometry_msgs::Point getPoseOfNextWaypoint() const
-  {
-    return current_waypoints_.getWaypointPosition(num_of_next_waypoint_);
-  }
-  geometry_msgs::Point getPoseOfNextTarget() const
-  {
-    return position_of_next_target_;
-  }
-  geometry_msgs::Pose getCurrentPose() const
-  {
-    return current_pose_.pose;
-  }
-
-  double getLookaheadDistance() const
-  {
-    return lookahead_distance_;
-  }
-  // processing for ROS
-  geometry_msgs::TwistStamped go();
 };
-}
+}  // waypoint_follower
 
-#endif  // PURE_PURSUIT_CORE_H
+#endif  // PURE_PURSUIT_H
