+++ ros/src/computing/planning/motion/packages/waypoint_follower/nodes/pure_pursuit/pure_pursuit_core.h
@@ -1,0 +1,90 @@
+/*
+ *  Copyright (c) 2015, Nagoya University
+ *  All rights reserved.
+ *
+ *  Redistribution and use in source and binary forms, with or without
+ *  modification, are permitted provided that the following conditions are met:
+ *
+ *  * Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer.
+ *
+ *  * Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ *  * Neither the name of Autoware nor the names of its
+ *    contributors may be used to endorse or promote products derived from
+ *    this software without specific prior written permission.
+ *
+ *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ *  AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ *  IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ *  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ *  FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ *  DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ *  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ *  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ *  OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ *  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+
+#ifndef PURE_PURSUIT_CORE_H
+#define PURE_PURSUIT_CORE_H
+
+// ROS includes
+#include <ros/ros.h>
+#include <geometry_msgs/TwistStamped.h>
+#include <geometry_msgs/PoseStamped.h>
+#include <visualization_msgs/Marker.h>
+
+// User defined includes
+#include "runtime_manager/ConfigWaypointFollower.h"
+#include "waypoint_follower/lane.h"
+#include "pure_pursuit_viz.h"
+#include "pure_pursuit.h"
+
+namespace waypoint_follower
+{
+class PurePursuitNode
+{
+public:
+  PurePursuitNode();
+  ~PurePursuitNode();
+
+  void run();
+
+private:
+  // handle
+  ros::NodeHandle nh_;
+  ros::NodeHandle private_nh_;
+
+  // class
+  PurePursuit pp_;
+
+  // publisher
+  ros::Publisher pub1_, pub2_, pub11_, pub12_, pub13_, pub14_, pub15_;
+
+  // subscriber
+  ros::Subscriber sub1_, sub2_, sub3_, sub4_;
+
+  // constant
+  const int LOOP_RATE_;  // processing frequency
+
+  // variables
+  bool is_linear_interpolation_;
+
+  // callbacks
+  void callbackFromConfig(const runtime_manager::ConfigWaypointFollowerConstPtr &config);
+  void callbackFromCurrentPose(const geometry_msgs::PoseStampedConstPtr &msg);
+  void callbackFromCurrentVelocity(const geometry_msgs::TwistStampedConstPtr &msg);
+  void callbackFromWayPoints(const waypoint_follower::laneConstPtr &msg);
+
+  // initializer
+  void initForROS();
+
+  // functions
+  void publish();
+};
+}  // waypoint_follower
+
+#endif  // PURE_PURSUIT_CORE_H
