+++ ros/src/computing/planning/motion/packages/waypoint_follower/nodes/pure_pursuit/pure_pursuit_viz.h
@@ -35,12 +35,15 @@
 #include <ros/ros.h>
 #include <geometry_msgs/TwistStamped.h>
 #include <geometry_msgs/PoseStamped.h>
-#include "runtime_manager/ConfigWaypointFollower.h"
+
 #include <visualization_msgs/Marker.h>
 
 // C++ includes
 #include <memory>
+
+// User defined includes
 #include "waypoint_follower/libwaypoint_follower.h"
+#include "runtime_manager/ConfigWaypointFollower.h"
 
 namespace waypoint_follower
 {
@@ -52,13 +55,13 @@ visualization_msgs::Marker displayNextTarget(geometry_msgs::Point target);
 double calcRadius(geometry_msgs::Point target, geometry_msgs::Pose current_pose);
 
 // generate the locus of pure pursuit
-std::vector<geometry_msgs::Point> generateTrajectoryCircle(geometry_msgs::Point target, geometry_msgs::Pose current_pose);
+std::vector<geometry_msgs::Point> generateTrajectoryCircle(geometry_msgs::Point target,
+                                                           geometry_msgs::Pose current_pose);
 // display the locus of pure pursuit by markers.
 visualization_msgs::Marker displayTrajectoryCircle(std::vector<geometry_msgs::Point> traj_circle_array);
 
 // display the search radius by markers.
 visualization_msgs::Marker displaySearchRadius(geometry_msgs::Point current_pose, double search_radius);
-
 }
 
 #endif  // PURE_PURSUIT_VIZ_H
